% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustexp.R
\name{adjustexp}
\alias{adjustexp}
\title{Title}
\usage{
adjustexp(hazard, tdist, nonburnable)
}
\arguments{
\item{hazard}{hazard}

\item{tdist}{tdist}

\item{nonburnable}{nonburnable}
}
\value{
SpatRaster
}
\description{
Title
}
\examples{
# coming soon

}
