% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direxp.R
\name{direxp}
\alias{direxp}
\title{Directional Exposure}
\usage{
direxp(expl, value, plot = F, map = F, table = F)
}
\arguments{
\item{expl}{Spatraster from exposure()}

\item{value}{Spatvector of value as a point or simplified polygon}

\item{plot}{logical. if true, returns a directional plot}

\item{map}{logical, if true, returns a map of the viable transects}

\item{table}{logical, if true, returns a table of viable transects}
}
\value{
a spatvector of transects unless plot, map, or table are set to TRUE
}
\description{
Directional Exposure
}
