% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{exposure}
\alias{exposure}
\title{Calculate Exposure}
\usage{
exposure(hazard, tdist = c("l", "s", "r"), nonburnable)
}
\arguments{
\item{hazard}{a SpatRaster that represents hazardous fuels for the
transmission distance specified in tdist}

\item{tdist}{a character vector, can be 'l' for long-range embers, 's' for
short-range embers, or 'r' for radiant heat}

\item{nonburnable}{(optional) a spatRaster that represents the burnable
landscape. Any cells that cannot receive wildfire (e.g. open water,
rocks) should be of value 1, all other cells should be NODATA}
}
\value{
a SpatRaster
}
\description{
Calculate Exposure
}
\examples{
lhaz <- terra::rast(system.file("extdata/LHazAB2020.tif", package = "fireexposuR"))
nonburnable <- terra::rast(system.file("extdata/nonburnableAB2020.tif", package = "fireexposuR"))
lexp <- exposure(lhaz, tdist = "l")
lexpburnable <- exposure(lhaz, tdist = "l", nonburnable)

}
