% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireexposuR-package.R
\docType{package}
\name{fireexposuR-package}
\alias{fireexposuR}
\alias{fireexposuR-package}
\title{fireexposuR: Compute and Visualize Wildfire Exposure}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Methods for computing and visualizing wildfire igntion exposure and directional vulnerability that are published in a series of scientific publications are automated by the functions in this package. See Beverly et al. (2010) \doi{10.1071/WF09071}, Beverly et al. (2021) \doi{10.1007/s10980-020-01173-8}, and Beverly and Forbes (2023) \doi{10.1007/s11069-023-05885-3} for background and methodology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/fireexposuR}
  \item \url{https://docs.ropensci.org/fireexposuR/}
  \item Report bugs at \url{https://github.com/ropensci/fireexposuR/issues}
}

}
\author{
\strong{Maintainer}: Air Forbes \email{heyairf@gmail.com} (\href{https://orcid.org/0000-0002-9842-7648}{ORCID})

Other contributors:
\itemize{
  \item Jennifer Beverly [thesis advisor]
}

}
\keyword{internal}
