% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapexpclass.R
\name{mapexpclass}
\alias{mapexpclass}
\title{Map exposure with a classified scale}
\usage{
mapexpclass(exposure, classify = c("local", "landscape"), aoi, zoom)
}
\arguments{
\item{exposure}{SpatRaster (e.g. from \code{\link[=exposure]{exposure()}})}

\item{classify}{character, either \code{"local"} or \code{"landscape"} to specify
classification scheme to use.}

\item{aoi}{SpatVector of an area of interest to mask exposure}

\item{zoom}{(Optional). numeric, set the zoom level for the basemap based
on the extent of your data if defaults are not appropriate. Defaults if:
\itemize{
\item \code{classify = "local"} the zoom level default is \code{13}
\item \code{classify = "landscape"} the zoom level default is \code{6}
}}
}
\value{
a standardized map is returned as a ggplot object
}
\description{
\code{mapexpclass()} produces a standardized map of exposure with
exposure classes. This function dynamically pulls map tiles for a base map,
so it is recommended the area of interest is localized. The zoom level may
need to be adjusted based on the extent of your data; see
\href{https://wiki.openstreetmap.org/wiki/Zoom_levels}{OpenStreetMap Wiki} for
more information on zoom levels.
For mapping large extents it is recommended (and will be faster) to use
\code{\link[=mapexpcont]{mapexpcont()}} which does not use base maps.

Scales and colors are determined by the parameter \code{classify}
which can be set to \code{"local"} or \code{"landscape"}.

Landscape classification breaks are:
\itemize{
\item Low (0-20\%)
\item Moderate (20-40\%)
\item High (40-60\%),
\item Very High (60-80\%)
\item Extreme (80-100\%)
}

Local classification breaks are:
\itemize{
\item Nil (0\%)
\item Low (>0-15\%)
\item Moderate (15-30\%)
\item High (30-45\%)
\item Extreme (45\%+)
}
}
\examples{
# generate example hazard data -----------------------------
set.seed(0)
e <- c(45,55,495,505) * 10000
r <- terra::rast(resolution = 100, extent = terra::ext(e))
terra::values(r) <- sample(c(0,1), terra::ncell(r), replace = TRUE)
terra::crs(r) <- "EPSG:32608"
r <- terra::sieve(r, threshold = 50, directions = 4)
haz <- terra::sieve(r, threshold = 500, directions = 4)
# generate example AOI polygon -----------------------------
filepath <- "extdata/builtsimpleexamplegeom.csv"
g <- read.csv(system.file(filepath, package = "fireexposuR"))
m <- as.matrix(g)
v <- terra::vect(m, "polygons", crs = haz)
# ----------------------------------------------------------

exp <- exposure(haz)
mapexpclass(exp, classify = "local", v)

}
\seealso{
\code{\link[=exposure]{exposure()}}, \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
