% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapexpcont.R
\name{mapexpcont}
\alias{mapexpcont}
\title{Map exposure with a continuous scale}
\usage{
mapexpcont(exposure, aoi)
}
\arguments{
\item{exposure}{SpatRaster from \code{\link[=exposure]{exposure()}}}

\item{aoi}{(optional) SpatVector of an area of interest to mask the exposure}
}
\value{
a map is returned as a ggplot object
}
\description{
\code{mapexpcont()} produces a standardized map of exposure with a
continuous scale. An optional area of interest can be used as a mask. The
ggplot object returned can be further modified with the ggplot library.
}
\examples{

#' # generate example hazard data -----------------------------
set.seed(0)
e <- c(45,55,495,505) * 10000
r <- terra::rast(resolution = 100, extent = terra::ext(e))
terra::values(r) <- sample(c(0,1), terra::ncell(r), replace = TRUE)
terra::crs(r) <- "EPSG:32608"
r <- terra::sieve(r, threshold = 50, directions = 4)
haz <- terra::sieve(r, threshold = 500, directions = 4)
# -------------------------------------------------------------

exp <- exposure(haz)
mapexpcont(exp)

}
\seealso{
\code{\link[=exposure]{exposure()}}, \code{\link[=ggplot]{ggplot()}}
}
