% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplandscape.R
\name{maplandscape}
\alias{maplandscape}
\title{Map continuous exposure across a landscape}
\usage{
maplandscape(exposure, aoi)
}
\arguments{
\item{exposure}{SpatRaster from [exposure()]}

\item{aoi}{(optional) SpatVector of an area of interest to mask exposure for
summary}
}
\value{
a map is returned as a ggplot object
}
\description{
`maplandscape()` is a helper function to produce a standardized
map of exposure across a landscape with a continuous scale. The ggplot object
returned can be further modified with the ggplot library.
}
\examples{

#' # generate example hazard data -----------------------------
set.seed(0)
e <- c(45,55,495,505) * 10000
r <- terra::rast(resolution = 100, extent = terra::ext(e))
terra::values(r) <- sample(c(0,1), terra::ncell(r), replace = TRUE)
terra::crs(r) <- "EPSG:32608"
r <- terra::sieve(r, threshold = 50, directions = 4)
haz <- terra::sieve(r, threshold = 500, directions = 4)
# -------------------------------------------------------------

exp <- exposure(haz)
maplandscape(exp)

# IGNORE BELOW
# filepath <- "extdata/LExpAB2020.tif"
# lexp <- terra::rast(system.file(filepath, package = "fireexposuR"))
# filepath <- "extdata/fpa.shp"
# fpa <- terra::vect(system.file(filepath, package = "fireexposuR"))

}
\seealso{
[exposure()], [ggplot()]
}
