% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplocal.R
\name{maplocal}
\alias{maplocal}
\title{Map classified exposure for a localized area}
\usage{
maplocal(exposure, aoi)
}
\arguments{
\item{exposure}{SpatRaster from [exposure()]}

\item{aoi}{SpatVector of an area of interest to mask exposure for summary}
}
\value{
a standardized map is returned as a ggplot object
}
\description{
`maplocal()` is a helper function to produce a standardized map
  of exposure within a localized area with exposure classes. The ggplot
  object returned can be further modified with the ggplot library.

  #' Local classification breaks are:
* Nil (0%)
* Low (>0-15%)
* Moderate (15-30%)
* High (30-45%)
* Extreme (45%+)
}
\examples{
# generate example hazard data -----------------------------
set.seed(0)
e <- c(45,55,495,505) * 10000
r <- terra::rast(resolution = 100, extent = terra::ext(e))
terra::values(r) <- sample(c(0,1), terra::ncell(r), replace = TRUE)
terra::crs(r) <- "EPSG:32608"
r <- terra::sieve(r, threshold = 50, directions = 4)
haz <- terra::sieve(r, threshold = 500, directions = 4)
# generate example AOI polygon -----------------------------
filepath <- "extdata/builtsimpleexamplegeom.csv"
g <- read.csv(system.file(filepath, package = "fireexposuR"))
m <- as.matrix(g)
v <- terra::vect(m, "polygons", crs = haz)
# ----------------------------------------------------------

exp <- exposure(haz)
maplocal(exp, v)

}
\seealso{
[exposure()], [ggplot()]
}
