% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summexp.R
\name{summexp}
\alias{summexp}
\title{Summarize exposure by class}
\usage{
summexp(exposure, aoi, classify = c("landscape", "local"))
}
\arguments{
\item{exposure}{SpatRaster from [exposure()]}

\item{aoi}{(optional) SpatVector of an area of interest to mask exposure for summary}

\item{classify}{character, either `"landscape"` or `"local"`. default is `"landscape"`.}
}
\value{
a summary table as a data frame object
}
\description{
`summexp()` creates a summary table of exposure by the user specified
classification scheme. The table reports the number of pixels, the
proportion, and area in hectares and meters squared by class.

Landscape classification breaks are:
* Low (0-20%)
* Moderate (20-40%)
* High (40-60%),
* Very High (60-80%)
* Extreme (80-100%)

Local classification breaks are:
* Nil (0%)
* Low (>0-15%)
* Moderate (15-30%)
* High (30-45%)
* Extreme (45%+)
}
\examples{

lexp <- terra::rast(system.file("extdata/LExpAB2020.tif", package = "fireexposuR"))
fpa <- terra::vect(system.file("extdata/fpa.shp", package = "fireexposuR"))
WHITpoly <- terra::vect(system.file("extdata/WHITpoly.shp", package = "fireexposuR"))

#summary table for landscape classification, entire extent
summexp(lexp, classify = "landscape")

#summary table for landscape classification, for forest protection area
summexp(lexp, fpa, classify = "landscape")

# summary table for local classification, built enviro of a community
summexp(lexp, WHITpoly, classify = "local")

}
\seealso{
[exposure()]
}
