% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateexp.R
\name{validateexp}
\alias{validateexp}
\title{DOESNT WORK YET :P}
\usage{
validateexp(burnableexposure, fires, sampleper = 0.001, aoi, plot = FALSE)
}
\arguments{
\item{burnableexposure}{exposure}

\item{fires}{fires}

\item{sampleper}{sample percent}

\item{aoi}{aoi}

\item{plot}{boolean}
}
\value{
table or plots
}
\description{
DOESNT WORK YET :P
}
\examples{
# coming soon?

}
