% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustexp.R
\name{adjustexp}
\alias{adjustexp}
\title{Compute the wildfire exposure metric with custom transmission distances}
\usage{
adjustexp(hazard, tdist, nonburnable)
}
\arguments{
\item{hazard}{a SpatRaster that represents hazardous fuels for the
transmission distance specified in \code{tdist}}

\item{tdist}{Numeric, transmission distance in meters}

\item{nonburnable}{(Optional) a SpatRaster that represents the burnable
landscape. Any cells that cannot receive wildfire (e.g. open water, rock)
should be of value 1, all other cells should be NODATA. This parameter
should be provided if preparing data for \code{\link[=validateexp]{validateexp()}}}
}
\value{
SpatRaster object of exposure values
}
\description{
For advanced users. If the transmission distances from the
wildfire exposure literature are not representative of the wildland fuels
in your area of interest, this function can be used to change the
transmission distance to a custom distance. It is highly recommended that
any exposure layers produced with this function are validated with observed
fire history using the \code{\link[=validateexp]{validateexp()}} function.
}
\examples{
#' # generate example hazard data -----------------------------
set.seed(0)
e <- c(45,55,495,505) * 10000
r <- terra::rast(resolution = 100, extent = terra::ext(e))
terra::values(r) <- sample(c(0,1), terra::ncell(r), replace = TRUE)
r <- terra::sieve(r, threshold = 50, directions = 4)
haz <- terra::sieve(r, threshold = 500, directions = 4)
# -----------------------------------------------------------

# compute long range exposure with custom disance of 800 m
exp <- adjustexp(haz, tdist = 800)
exp

}
