% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direxp.R
\name{direxp}
\alias{direxp}
\title{Directional Exposure
exposu}
\usage{
direxp(exposure, value, plot = FALSE, map = FALSE, table = FALSE)
}
\arguments{
\item{exposure}{Spatraster from exposure()}

\item{value}{Spatvector of value as a point or simplified polygon}

\item{plot}{logical. if true, returns a directional plot}

\item{map}{logical, if true, returns a map of the viable transects}

\item{table}{logical, if true, returns a table of viable transects}
}
\value{
a spatvector of transects unless plot, map, or table are set to TRUE
}
\description{
Directional Exposure
exposu
}
\examples{

lexp <- terra::rast(system.file("extdata/LExpAB2020.tif", package = "fireexposuR"))
WHITpoly <- terra::vect(system.file("extdata/WHITpoly.shp", package = "fireexposuR"))
WHITpt <- terra::centroids(WHITpoly)

# return a SpatVector that can be used in R or exported
direxp(lexp, WHITpoly)

# generate a map in R using a polygon
direxp(lexp, WHITpoly, map = TRUE)

# generate a map in R using a point
direxp(lexp, WHITpt, map = TRUE)

# or generate plots
direxp(lexp, WHITpoly, plot = TRUE)
direxp(lexp, WHITpt, plot = TRUE)
}
