% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{exposure}
\alias{exposure}
\title{Calculate Exposure}
\usage{
exposure(haz, tdist = c("l", "s", "r"), nonburnable)
}
\arguments{
\item{haz}{a SpatRaster that represents hazardous fuels for the transmission
distance specified in tdist}

\item{tdist}{a character vector, can be 'l' for long-range embers, 's' for
short-range embers, or 'r' for radiant heat}

\item{nonburnable}{(optional) a spatRaster that represents the burnable landscape. Any cells
that cannot receive wildfire (e.g. open water, rocks) should be of value 1,
all other cells should be of value 0}
}
\value{
a SpatRaster
}
\description{
Calculate Exposure
}
