% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractexp.R
\name{extractexp}
\alias{extractexp}
\title{Title}
\usage{
extractexp(
  exposure,
  values,
  method = c("max", "mean"),
  classify = c("landscape", "local"),
  summary = FALSE,
  map = FALSE
)
}
\arguments{
\item{exposure}{spatraster from exposure()}

\item{values}{spatvector of points or polygons}

\item{method}{character, either "max" or "mean". required when values are
polygons}

\item{classify}{character, either "local" or "landscape" to specify
classification style}

\item{summary}{logical, if True returns a summary table}

\item{map}{logical, if true returns a ggplot map of values coloured by
classification}
}
\value{
spatvector unless summary or map = T
}
\description{
Title
}
\examples{

lexp <- terra::rast(system.file("extdata/LExpAB2020.tif", package = "fireexposuR"))
WHITstruc <- terra::vect(system.file("extdata/WHITstruc.shp", package = "fireexposuR"))

# summary table of structures in each class (local classification)
extractexp(lexp, WHITstruc, method = "mean", classify = "local", summary = TRUE)

# map structures in each class (local classification)
extractexp(lexp, WHITstruc, method = "mean", classify = "local", map = TRUE)

}
