% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractexp.R
\name{extractexp}
\alias{extractexp}
\title{Title}
\usage{
extractexp(
  exp,
  values,
  method = c("max", "mean"),
  classify = c("landscape", "local"),
  summary = F,
  map = F
)
}
\arguments{
\item{exp}{spatraster from exposure()}

\item{values}{spatvector of points or polygons}

\item{method}{character, either "max" or "mean". required when values are polygons}

\item{classify}{character, either "local" or "landscape" to specify classification style}

\item{summary}{logical, if True returns a summary table}

\item{map}{logical, if true returns a ggplot map of values coloured by classification}
}
\description{
Title
}
