% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_exp_dir.R
\name{fire_exp_dir}
\alias{fire_exp_dir}
\title{Conduct a directional exposure assessment}
\usage{
fire_exp_dir(exposure, value, table = FALSE)
}
\arguments{
\item{exposure}{SpatRaster (e.g. from \code{\link[=fire_exp]{fire_exp()}})}

\item{value}{Spatvector of value as a point or simplified polygon}

\item{table}{Boolean, when \code{TRUE}: returns a table instead of a feature. The
default is \code{FALSE}.}
}
\value{
a SpatVector of the transects with a attributes: degree, segment,
viable. Unless:
\itemize{
\item \code{table = TRUE}: a data frame
}
}
\description{
\code{fire_exp_dir()} automates the directional vulnerability assessment
methods from Beverly and Forbes 2023. This function can return the
assessment transects as a spatial feature of transect segments or a table
of each transect segment as a data frame
The exposure layer and value layer must have a CRS defined.
If a polygon is used, it must be a simple feature without holes or complex
geometry. The function will not run if the polygon should be simplified
further. See Forbes and Beverly 2024 (Manuscript in preparation) for data
preparation advice.
}
\examples{
# read example hazard data ----------------------------------
filepath <- "extdata/hazard.tif"
haz <- terra::rast(system.file(filepath, package = "fireexposuR"))
# generate an example point ---------------------------------
wkt <- "POINT (500000 5000000)"
pt <- terra::vect(wkt, crs = haz)
# -----------------------------------------------------------

exp <- fire_exp(haz)

fire_exp_dir(exp, pt)

}
