% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_exp_dir_plot.R
\name{fire_exp_dir_plot}
\alias{fire_exp_dir_plot}
\title{visualize directional exposure}
\usage{
fire_exp_dir_plot(transects, value, map = FALSE)
}
\arguments{
\item{transects}{SpatVector (output from \code{\link[=fire_exp_dir]{fire_exp_dir()}})}

\item{value}{the same Spatvector of value as a point or simplified polygon
used for \code{\link[=fire_exp_dir]{fire_exp_dir()}}}

\item{map}{Boolean, when \code{TRUE}: returns a map of the viable transects. The
default is \code{FALSE}}
}
\value{
a standardized plot as a ggplot object, or if \code{map = TRUE} a
standardized map as a ggplot object
}
\description{
\code{fire_exp_dir_plot()} plots directional exposure following
methods from Beverly and Forbes 2023 as a standardized radial plot,
mapped with a basemap, or as a table.
}
\examples{
# read example hazard data ----------------------------------
filepath <- "extdata/hazard.tif"
haz <- terra::rast(system.file(filepath, package = "fireexposuR"))
# generate an example point ---------------------------------
wkt <- "POINT (500000 5000000)"
pt <- terra::vect(wkt, crs = haz)
# -----------------------------------------------------------

exp <- fire_exp(haz)

# compute transects
transects <- fire_exp_dir(exp, pt)

# radial plot of directional exposure
fire_exp_dir_plot(transects, pt)

#' # mapped directional exposure transects
# note: example data is drawn in the ocean so basemap is not representative
fire_exp_dir_plot(transects, pt, map = TRUE)
}
