% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_exp_extract.R
\name{fire_exp_extract}
\alias{fire_exp_extract}
\title{Extract exposure values to features}
\usage{
fire_exp_extract(exposure, values)
}
\arguments{
\item{exposure}{SpatRaster (e.g. from \code{\link[=fire_exp]{fire_exp()}})}

\item{values}{Spatvector of points or polygons}
}
\value{
a SpatVector object with new attributes
}
\description{
\code{fire_exp_extract()} extracts the underlying exposure value for
each feature in the values layer.
}
\details{
\strong{DOCUMENTATION IN DEVELOPMENT}
Values can be provided as either points or
polygons, and must be singlepart features (i.e. the attribute table has one
row per value). If the values are polygon features both the maximum and mean
exposure is computed.
}
\examples{
# read example hazard data
hazard_file_path <- "extdata/hazard.tif"
hazard <- terra::rast(system.file(hazard_file_path, package = "fireexposuR"))

# generate example area of interest geometry
geom_file_path <- "extdata/polygon_geometry.csv"
geom <- read.csv(system.file(geom_file_path, package = "fireexposuR"))
aoi <- terra::vect(as.matrix(geom), "polygons", crs = hazard)

# generate random points within the aoi polygon
points <- terra::spatSample(aoi, 100)

# compute exposure
exposure <- fire_exp(hazard)

fire_exp_extract(exposure, points)

}
