% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_exp_extract.R
\name{fire_exp_extract}
\alias{fire_exp_extract}
\title{Extract exposure values to features}
\usage{
fire_exp_extract(exposure, values)
}
\arguments{
\item{exposure}{SpatRaster (e.g. from \code{\link[=fire_exp]{fire_exp()}})}

\item{values}{Spatvector of points or polygons}
}
\value{
a SpatVector object with new attributes
}
\description{
\code{fire_exp_extract()} extracts the underlying exposure value for
each feature in the values layer. Values can be provided as either points or
polygons, and must be singlepart features (i.e. the attribute table has one
row per value). If the values are polygon features both the maximum and mean
exposure is computed.
}
\examples{
# read example hazard data ----------------------------------
filepath <- "extdata/hazard.tif"
haz <- terra::rast(system.file(filepath, package = "fireexposuR"))
# read example AOI
filepath <- "extdata/builtsimpleexamplegeom.csv"
g <- read.csv(system.file(filepath, package = "fireexposuR"))
v <- terra::vect(as.matrix(g), "polygons", crs = haz)
# generate random points
pts <- terra::spatSample(v, 200)
# ----------------------------------------------------------

exp <- fire_exp(haz)

# extract exposure to point values
fire_exp_extract(exp, pts)

}
