% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_exp_map_class.R
\name{fire_exp_map_class}
\alias{fire_exp_map_class}
\title{Map exposure with a classified scale}
\usage{
fire_exp_map_class(exposure, classify = c("local", "landscape"), aoi, zoom)
}
\arguments{
\item{exposure}{SpatRaster (e.g. from \code{\link[=fire_exp]{fire_exp()}})}

\item{classify}{character, either \code{"local"} or \code{"landscape"} to specify
classification scheme to use.}

\item{aoi}{SpatVector of an area of interest to mask exposure}

\item{zoom}{(Optional). numeric, set the zoom level for the basemap based
on the extent of your data if defaults are not appropriate. Defaults if:
\itemize{
\item \code{classify = "local"} the zoom level default is \code{13}
\item \code{classify = "landscape"} the zoom level default is \code{6}
}}
}
\value{
a standardized map is returned as a ggplot object
}
\description{
\code{fire_exp_map_class()} produces a standardized map by
classifying exposure into predetermined exposure classes.
}
\details{
\strong{DOCUMENTATION IN DEVELOPMENT}
This function dynamically pulls map tiles for a base map,
so it is recommended the area of interest is localized. The zoom level may
need to be adjusted based on the extent of your data; see
\href{https://wiki.openstreetmap.org/wiki/Zoom_levels}{OpenStreetMap Wiki} for
more information on zoom levels.
For mapping large extents it is recommended (and will be faster) to use
\code{\link[=fire_exp_map_cont]{fire_exp_map_cont()}} which does not use base maps.

Scales and colors are determined by the parameter \code{classify}
which can be set to \code{"local"} or \code{"landscape"}.

Landscape classification breaks are:
\itemize{
\item Low (0-20\%)
\item Moderate (20-40\%)
\item High (40-60\%),
\item Very High (60-80\%)
\item Extreme (80-100\%)
}

Local classification breaks are:
\itemize{
\item Nil (0\%)
\item Low (>0-15\%)
\item Moderate (15-30\%)
\item High (30-45\%)
\item Extreme (45\%+)
}
}
\examples{
# read example hazard data
hazard_file_path <- "extdata/hazard.tif"
hazard <- terra::rast(system.file(hazard_file_path, package = "fireexposuR"))

# generate example area of interest geometry
geom_file_path <- "extdata/polygon_geometry.csv"
geom <- read.csv(system.file(geom_file_path, package = "fireexposuR"))
aoi <- terra::vect(as.matrix(geom), "polygons", crs = hazard)

# compute exposure
exposure <- fire_exp(hazard)

fire_exp_map_class(exposure, classify = "local", aoi)

}
