% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_exp_summary.R
\name{fire_exp_summary}
\alias{fire_exp_summary}
\title{Summarize exposure by class}
\usage{
fire_exp_summary(exposure, aoi, classify = c("landscape", "local"))
}
\arguments{
\item{exposure}{SpatRaster from \code{\link[=fire_exp]{fire_exp()}}}

\item{aoi}{(optional) SpatVector of an area of interest to mask exposure for
summary}

\item{classify}{character, either \code{"landscape"} or \code{"local"}. default is
\code{"landscape"}.}
}
\value{
a summary table as a data frame object
}
\description{
\code{fire_exp_summary()} creates a summary table of exposure by the
specified classification scheme. The table reports the number of pixels, the
proportion, and area in hectares and meters squared by class.

Landscape classification breaks are:
\itemize{
\item Low (0-20\%)
\item Moderate (20-40\%)
\item High (40-60\%),
\item Very High (60-80\%)
\item Extreme (80-100\%)
}

Local classification breaks are:
\itemize{
\item Nil (0\%)
\item Low (>0-15\%)
\item Moderate (15-30\%)
\item High (30-45\%)
\item Extreme (45\%+)
}
}
\examples{
#' # read example hazard data ----------------------------------
filepath <- "extdata/hazard.tif"
haz <- terra::rast(system.file(filepath, package = "fireexposuR"))
# generate example AOI polygon -----------------------------
filepath <- "extdata/builtsimpleexamplegeom.csv"
g <- read.csv(system.file(filepath, package = "fireexposuR"))
v <- terra::vect(as.matrix(g), "polygons", crs = haz)
# ----------------------------------------------------------

exp <- fire_exp(haz)

# for full extent of data
fire_exp_summary(exp, classify = "landscape")

# for a masked area
fire_exp_summary(exp, v, classify = "landscape")

}
