% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_exp_summary.R
\name{fire_exp_summary}
\alias{fire_exp_summary}
\title{Summarize exposure by class}
\usage{
fire_exp_summary(exposure, aoi, classify = c("landscape", "local"))
}
\arguments{
\item{exposure}{SpatRaster from \code{\link[=fire_exp]{fire_exp()}}}

\item{aoi}{(optional) SpatVector of an area of interest to mask exposure for
summary}

\item{classify}{character, either \code{"landscape"} or \code{"local"}. default is
\code{"landscape"}.}
}
\value{
a summary table as a data frame object
}
\description{
\code{fire_exp_summary()} creates a summary table of area and
proportions of exposure in predetermined exposure classes.
}
\details{
\strong{DOCUMENTATION IN DEVELOPMENT}

The table reports the number of pixels, the
proportion, and area in hectares and meters squared by class.

Landscape classification breaks are:
\itemize{
\item Low (0-20\%)
\item Moderate (20-40\%)
\item High (40-60\%),
\item Very High (60-80\%)
\item Extreme (80-100\%)
}

Local classification breaks are:
\itemize{
\item Nil (0\%)
\item Low (>0-15\%)
\item Moderate (15-30\%)
\item High (30-45\%)
\item Extreme (45\%+)
}
}
\examples{
# read example hazard data
hazard_file_path <- "extdata/hazard.tif"
hazard <- terra::rast(system.file(hazard_file_path, package = "fireexposuR"))

# read example area of interest polygon geometry
geom_file_path <- "extdata/polygon_geometry.csv"
geom <- read.csv(system.file(geom_file_path, package = "fireexposuR"))
aoi <- terra::vect(as.matrix(geom), "polygons", crs = hazard)

# Compute exposure
exposure <- fire_exp(hazard)

# Summary for full extent of data
fire_exp_summary(exposure, classify = "landscape")

# Summary masked to an area of interest
fire_exp_summary(exposure, aoi, classify = "landscape")

}
