% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_exp_validate.R
\name{fire_exp_validate}
\alias{fire_exp_validate}
\title{Validate exposure with observed fires}
\usage{
fire_exp_validate(
  burnableexposure,
  fires,
  aoi,
  samplesize = 0.005,
  plot = FALSE
)
}
\arguments{
\item{burnableexposure}{A SpatRaster of exposure, non-burnable cells should
be removed using optional parameter in \code{\link[=fire_exp]{fire_exp()}}.}

\item{fires}{A SpatVector of observed fire perimeters}

\item{aoi}{(Optional) A SpatVector that delineates an area of interest}

\item{samplesize}{Proportion of area to sample. The default is \code{0.005} (0.5\%)}

\item{plot}{Boolean, when \code{TRUE}: returns a plot of expected and observed
exposure class proportions within entire extent and sampled areas. The
default is \code{FALSE}.}
}
\value{
table or plots
}
\description{
For advanced users. \code{fire_exp_validate()} compares the
proportion of exposure classes in a the study area to the proportion of
exposure classes within burned areas. A random sample is taken to account
for spatial autocorrelation.
}
\examples{

# read example hazard data ----------------------------------
filepath <- "extdata/hazard.tif"
haz <- terra::rast(system.file(filepath, package = "fireexposuR"))
# generate example non-burn data -----------------------------
filepath <- "extdata/builtsimpleexamplegeom.csv"
g <- read.csv(system.file(filepath, package = "fireexposuR"))
v <- terra::vect(as.matrix(g), "polygons", crs = haz)
nb <- terra::rasterize(v, haz)
# generate example fire polygons -----------------------------
pts <- terra::spatSample(terra::rescale(haz, 0.8), 30, as.points = TRUE)
fires <- terra::buffer(pts, 800)
# ----------------------------------------------------------
# PLEASE NOTE THIS RANDOMLY GENERATED DATA DOES NOT GIVE MEANINGFUL RESULTS
exp <- fire_exp(haz, nonburnable = nb)
fire_exp_validate(exp, fires)
#' fire_exp_validate(exp, fires, plot = TRUE)

}
