% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_exp_validate.R
\name{fire_exp_validate}
\alias{fire_exp_validate}
\title{Validate exposure with observed fires}
\usage{
fire_exp_validate(
  burnableexposure,
  fires,
  aoi,
  samplesize = 0.005,
  plot = FALSE
)
}
\arguments{
\item{burnableexposure}{A SpatRaster of exposure, non-burnable cells should
be removed using optional parameter in \code{\link[=fire_exp]{fire_exp()}}.}

\item{fires}{A SpatVector of observed fire perimeters}

\item{aoi}{(Optional) A SpatVector that delineates an area of interest}

\item{samplesize}{Proportion of area to sample. The default is \code{0.005} (0.5\%)}

\item{plot}{Boolean, when \code{TRUE}: returns a plot of expected and observed
exposure class proportions within entire extent and sampled areas. The
default is \code{FALSE}.}
}
\value{
table or plot
}
\description{
For advanced users. \code{fire_exp_validate()} compares the
proportion of exposure classes in a the study area to the proportion of
exposure classes within burned areas. A random sample is taken to account
for spatial autocorrelation.
}
\details{
\strong{DOCUMENTATION IN DEVELOPMENT}
}
\examples{
# read example hazard data
hazard_file_path <- "extdata/hazard.tif"
hazard <- terra::rast(system.file(hazard_file_path, package = "fireexposuR"))

# generate example non-burnable cells data
geom_file_path <- "extdata/polygon_geometry.csv"
geom <- read.csv(system.file(geom_file_path, package = "fireexposuR"))
polygon <- terra::vect(as.matrix(geom), "polygons", crs = hazard)
no_burn <- terra::rasterize(polygon, hazard)

# generate example fire polygons by buffering random points
points <- terra::spatSample(terra::rescale(hazard, 0.8),
                            30, as.points = TRUE)
fires <- terra::buffer(points, 800)
# PLEASE NOTE THIS RANDOMLY GENERATED DATA DOES NOT GIVE MEANINGFUL RESULTS

# compute exposure and remove non-burnable cells
exposure <- fire_exp(hazard, nonburnable = no_burn)

# results as table
fire_exp_validate(exposure, fires)

# results as bar chart
fire_exp_validate(exposure, fires, plot = TRUE)

}
