% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireexposuR-package.R
\docType{package}
\name{fireexposuR-package}
\alias{fireexposuR}
\alias{fireexposuR-package}
\title{fireexposuR: Compute and Visualize Wildfire Exposure}
\description{
This package computes wildfire exposure using methods from Beverly et al. (2010), Beverly et al. (2021), and Beverly and Forbes (2023). It provides functions to standardize the mapping and visualization of wildfire exposure. This package requires pre-processing of data which is docomented in Forbes and Beverly (in preparation).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heyairf/fireexposuR}
  \item Report bugs at \url{https://github.com/heyairf/fireexposuR/issues}
}

}
\author{
\strong{Maintainer}: Air Forbes \email{amforbes@ualberta.ca} (\href{https://orcid.org/0000-0002-9842-7648}{ORCID})

Authors:
\itemize{
  \item Jennifer Beverly
}

}
\keyword{internal}
