% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireexposuR-package.R
\docType{package}
\name{fireexposuR-package}
\alias{fireexposuR}
\alias{fireexposuR-package}
\title{fireexposuR: Compute and Visualize Wildfire Exposure}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package computes and visualizes wildfire exposure using the methods documented in a series of scientific publications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heyairf/fireexposuR}
  \item \url{https://heyairf.github.io/fireexposuR/}
  \item Report bugs at \url{https://github.com/heyairf/fireexposuR/issues}
}

}
\author{
\strong{Maintainer}: Air Forbes \email{amforbes@ualberta.ca} (\href{https://orcid.org/0000-0002-9842-7648}{ORCID})

Other contributors:
\itemize{
  \item Jennifer Beverly [thesis advisor]
}

}
\keyword{internal}
