% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplandscape.R
\name{maplandscape}
\alias{maplandscape}
\title{Map continuous exposure across a landscape}
\usage{
maplandscape(exposure, aoi)
}
\arguments{
\item{exposure}{SpatRaster from [exposure()]}

\item{aoi}{(optional) SpatVector of an area of interest to mask exposure for
summary}
}
\value{
a map is returned as a ggplot object
}
\description{
`maplandscape()` is a helper function to produce a standardized
map of exposure across a landscape with a continuous scale. The ggplot object
returned can be further modified with the ggplot library.
}
\examples{
filepath <- "extdata/LExpAB2020.tif"
lexp <- terra::rast(system.file(filepath, package = "fireexposuR"))
filepath <- "extdata/fpa.shp"
fpa <- terra::vect(system.file(filepath, package = "fireexposuR"))

maplandscape(lexp, fpa)



}
\seealso{
[exposure()], [ggplot()]
}
