% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplocal.R
\name{maplocal}
\alias{maplocal}
\title{Map classified exposure for a localized area}
\usage{
maplocal(exposure, aoi)
}
\arguments{
\item{exposure}{SpatRaster from [exposure()]}

\item{aoi}{SpatVector of an area of interest to mask exposure for summary}
}
\value{
a standardized map is returned as a ggplot object
}
\description{
`maplocal()` is a helper function to produce a standardized map of
exposure within a localized area with exposure classes. The ggplot object
returned can be further modified with the ggplot library.

#' Local classification breaks are:
* Nil (0%)
* Low (>0-15%)
* Moderate (15-30%)
* High (30-45%)
* Extreme (45%+)
}
\examples{
lexp <- terra::rast(system.file("extdata/LExpAB2020.tif", package = "fireexposuR"))
WHITbuilt <- terra::vect(system.file("extdata/WHITbuilt.shp", package = "fireexposuR"))

maplocal(lexp, WHITbuilt)

}
\seealso{
[exposure()], [ggplot()]
}
