% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidirexp.R
\name{multidirexp}
\alias{multidirexp}
\title{Generate a directional load plot or table for multiple values}
\usage{
multidirexp(exposure, values, plot = FALSE, all = FALSE)
}
\arguments{
\item{exposure}{SpatRaster from [exposure()]}

\item{values}{Spatvector of value as a point or simplified polygon}

\item{plot}{Boolean, when `TRUE`: returns a standardized directional plot.
The default is `FALSE`.}

\item{all}{Boolean, when `TRUE`: considers all 3 segments (0-15km) of
directional transects. when `FALSE`: only the segments from 5-15 km are
included (Default)}
}
\value{
a SpatVector of the transects with a attributes: degree, segment,
  viable. Unless:
     * `plot = TRUE`: a standardized plot as a ggplot object
     * `map = TRUE`: a standardized map as a ggplot object
     * `table = TRUE`: a data frame with attributes by transect segment
}
\description{
`direxp()` automates the directional vulnerability assessment
methods from Beverly and Forbes 2023. This function can return directional
loads as:
 * a standardized radial plot as a ggplot object
 * a table summarizing if each degree is included by feature
}
\examples{
#coming soon
}
