% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidirexp.R
\name{multidirexp}
\alias{multidirexp}
\title{Generate a directional load plot or table for multiple values}
\usage{
multidirexp(exposure, values, plot = FALSE, all = FALSE)
}
\arguments{
\item{exposure}{SpatRaster from \code{\link[=exposure]{exposure()}}}

\item{values}{Spatvector of value as a point or simplified polygon}

\item{plot}{Boolean, when \code{TRUE}: returns a standardized directional plot.
The default is \code{FALSE}.}

\item{all}{Boolean, when \code{TRUE}: considers all 3 segments (0-15km) of
directional transects. when \code{FALSE}: only the segments from 5-15 km are
included (Default)}
}
\value{
a data.frame of the features with attributes: value featureID,
degree, to5 (binary), to10(binary), t015(binary), full(binary),
outer (binary). Unless:
* \code{plot = TRUE}: a standardized plot as a ggplot object
}
\description{
\code{multidirexp()} automates the directional vulnerability
assessment methods from Beverly and Forbes 2023. This function can return
directional loads as:
\itemize{
\item a standardized radial plot as a ggplot object
\item a table summarizing if each degree is included by feature
}
}
\examples{
# generate example hazard data -----------------------------
set.seed(0)
e <- c(45, 55, 495, 505) * 10000
r <- terra::rast(resolution = 100, extent = terra::ext(e))
terra::values(r) <- sample(c(0, 1), terra::ncell(r), replace = TRUE)
terra::crs(r) <- "EPSG:32608"
r <- terra::sieve(r, threshold = 50, directions = 4)
haz <- terra::sieve(r, threshold = 500, directions = 4)
# example points across the landscape ----------------------
e <- terra::buffer(terra::vect(terra::ext(haz), crs = haz), -15500)
pts <- terra::spatSample(e, 200)
# ----------------------------------------------------------

exp <- exposure(haz, tdist = "l")
# this example will take a while to run
\dontrun{
multidirexp(exp, pts, plot = TRUE)
}
}
