% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summexp.R
\name{summexp}
\alias{summexp}
\title{Title}
\usage{
summexp(exp, aoi, classify = c("landscape", "local"))
}
\arguments{
\item{exp}{spat raster from exposure()}

\item{aoi}{spatvector to mask exposure to for summaries}

\item{classify}{character, either "landscape" or "local"}
}
\value{
ggplot object
}
\description{
Title
}
