% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summexp.R
\name{summexp}
\alias{summexp}
\title{Summarize exposure by class}
\usage{
summexp(exposure, aoi, classify = c("landscape", "local"))
}
\arguments{
\item{exposure}{SpatRaster from [exposure()]}

\item{aoi}{(optional) SpatVector of an area of interest to mask exposure for
summary}

\item{classify}{character, either `"landscape"` or `"local"`. default is
`"landscape"`.}
}
\value{
a summary table as a data frame object
}
\description{
`summexp()` creates a summary table of exposure by the user
specified classification scheme. The table reports the number of pixels, the
proportion, and area in hectares and meters squared by class.

Landscape classification breaks are:
* Low (0-20%)
* Moderate (20-40%)
* High (40-60%),
* Very High (60-80%)
* Extreme (80-100%)

Local classification breaks are:
* Nil (0%)
* Low (>0-15%)
* Moderate (15-30%)
* High (30-45%)
* Extreme (45%+)
}
\examples{
# generate example hazard data -----------------------------
set.seed(0)
e <- c(45,55,495,505) * 10000
r <- terra::rast(resolution = 100, extent = terra::ext(e))
terra::values(r) <- sample(c(0,1), terra::ncell(r), replace = TRUE)
terra::crs(r) <- "EPSG:32608"
r <- terra::sieve(r, threshold = 50, directions = 4)
haz <- terra::sieve(r, threshold = 500, directions = 4)
# generate example AOI polygon -----------------------------
filepath <- "extdata/builtsimpleexamplegeom.csv"
g <- read.csv(system.file(filepath, package = "fireexposuR"))
m <- as.matrix(g)
v <- terra::vect(m, "polygons", crs = haz)
# ----------------------------------------------------------

exp <- exposure(haz)

# for full extent of data
summexp(exp, classify = "landscape")

# for a masked area
summexp(exp, v, classify = "landscape")

}
\seealso{
[exposure()]
}
